+++ instrumentation-genai/opentelemetry-instrumentation-vertexai/tests/test_chat_completions_experimental.py
@@ -358,7 +358,6 @@ def test_generate_content_all_events(
                 "You are a clever language model"
             ),
         ),
-        generate_content,
         log_exporter,
         instrument_with_experimental_semconvs,
     )
@@ -377,7 +376,6 @@ def test_preview_generate_content_all_input_events(
                 "You are a clever language model"
             ),
         ),
-        generate_content,
         log_exporter,
         instrument_with_experimental_semconvs,
     )
@@ -385,7 +383,6 @@ def test_preview_generate_content_all_input_events(
 
 def generate_content_all_input_events(
     model: GenerativeModel | PreviewGenerativeModel,
-    generate_content: callable,
     log_exporter: InMemoryLogExporter,
     instrument_with_experimental_semconvs: VertexAIInstrumentor,
 ):
